Move select case handlers from Movement to Chart
Moves code to handle selection of a specific trend case
from Movement classes to Chart classes.
