+++ src/market_analy/trends/movements.py
@@ -9,19 +9,16 @@ from dataclasses import dataclass
 from functools import cached_property
 
 import bqplot as bq
-import ipywidgets as w
 import numpy as np
 import pandas as pd
 
 from ..cases import CaseBase, CasesBase, CasesSupportsChartAnaly, CaseSupportsChartAnaly
-from ..charts import TOOLTIP_STYLE, Groups, tooltip_html_style
-from ..config import COL_ADV, COL_DEC
+from ..charts import tooltip_html_style
+from ..config import COL_ADV
 from ..formatters import formatter_datetime, formatter_float, formatter_percent
-from ..utils import bq_utils as ubq
 
 if typing.TYPE_CHECKING:
-    from .analy import TrendParams
-    from .charts import OHLCTrends
+    from .analy import TrendParams, TrendAltParams
 
 
 @dataclass(frozen=True, eq=False)
@@ -240,7 +237,7 @@ class MovementAlt(MovementBase, CaseSupportsChartAnaly):
         the reversal limit on the bar the movement start was confirmed.
     """
 
-    params: TrendParams
+    params: TrendAltParams
     sel: pd.Series
     start_conf_line: pd.Series
     eel: pd.Series | None
@@ -410,22 +407,6 @@ class MovementsSupportChartAnaly(CasesSupportsChartAnaly, typing.Protocol):
     @property
     def ends_conf_dec_px(self) -> list[float]:
         """Prices when end of decline movements confirmed."""
-        ...
-
-    def handler_hover_start(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing a movement start."""
-        ...
-
-    def handler_hover_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering over mark representing a movement end."""
-        ...
-
-    def handler_hover_conf_start(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering over mark representing when movement start confirmed."""
-        ...
-
-    def handler_hover_conf_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering over mark representing when movement end confirmed."""
         ...
 
     @property
@@ -618,22 +599,16 @@ class Movements(MovementsBase, MovementsSupportChartAnaly):
 
     cases: Sequence[Movement]
 
-    MARKER_MAP: typing.ClassVar = {
-        "cross": "End",
-        "circle": "Start conf",
-        "square": "End conf",
-    }
-
     def event_to_case(self, mark: bq.Scatter, event: dict) -> Movement:
         # Only for type checker to know move is a Movement as defined on this module.
         case = super().event_to_case(mark, event)
-        if typing.TYPE_CHECKING:
-            assert isinstance(case, Movement)
+        case = typing.cast(Movement, case)
         return case
 
     @staticmethod
-    def get_move_html(case: Movement) -> str:
+    def get_case_html(case: CaseSupportsChartAnaly) -> str:
         """Return html to describe a movement."""
+        case = typing.cast(Movement, case)
         color = "crimson" if case.is_dec else "limegreen"
         style = tooltip_html_style(color=color, line_height=1.3)
         s = f"<p {style}>Start: " + formatter_datetime(case.start)
@@ -653,130 +628,14 @@ class Movements(MovementsBase, MovementsSupportChartAnaly):
         if case.open:
             s += "None"
         else:
-            chg_color = "crimson" if case.conf_chg_pct < 0 else "limegreen"  # type: ignore
-            chg_style = tooltip_html_style(color=chg_color, line_height=1.3)
             if typing.TYPE_CHECKING:
                 assert case.conf_chg_pct is not None
+            chg_color = "crimson" if case.conf_chg_pct < 0 else "limegreen"
+            chg_style = tooltip_html_style(color=chg_color, line_height=1.3)
             s += f"<span {chg_style}>{formatter_percent(case.conf_chg_pct)}</span>"
 
         s += "</p"
         return s
-
-    def handler_hover_start(self, mark: bq.marks.Scatter, event: dict):
-        """Handler for hovering on mark representing a movement start."""
-        case = self.event_to_case(mark, event)
-        mark.tooltip.value = self.get_move_html(case)
-
-    def _handler_not_start(self, mark: bq.Scatter, event: dict):
-        """Handler to hover over any scatter mark not representing a movement start.
-
-        Displays tooltip describing what scatter represents and corresponding timestamp.
-        """
-        style = tooltip_html_style(color=mark.colors[0], line_height=1.3)
-        date = ubq.discontinuous_date_to_timestamp(event["data"]["x"])
-        name = self.MARKER_MAP[mark.marker]
-        s = f"<p {style}>{name}: {formatter_datetime(date)}"
-        s += f"<br>{name + ' px'}: {formatter_float(event['data']['y'])}</p"
-        mark.tooltip.value = s
-
-    def handler_hover_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing a movement end."""
-        self._handler_not_start(mark, event)
-
-    def handler_hover_conf_start(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing when movement start confirmed."""
-        self._handler_not_start(mark, event)
-
-    def handler_hover_conf_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing when movement end confirmed."""
-        self._handler_not_start(mark, event)
-
-    def handler_click_case(self, chart: OHLCTrends, mark: bq.Scatter, event: dict):
-        """Handler for clicking mark representing a movement start.
-
-        Removes all existing scatters from figure.
-
-        Adds scatters and lines for the single trend represented by the
-        clicked scatter point.
-        """
-        case = self.event_to_case(mark, event)
-        color = mark.colors[0]
-
-        marks = []
-
-        def f(*args):
-            marks.append(chart.create_scatter(*args))
-
-        def fl(data: pd.Series, color_: str, line_style: str, desc: str) -> bq.Lines:
-            tooltip_str = f"<p {tooltip_html_style(color=color_)}>{desc}</p>"
-            line = bq.Lines(
-                x=data.index,
-                y=data,
-                scales=chart.scales,
-                colors=[color_],
-                line_style=line_style,
-                tooltip=w.HTML(value=tooltip_str),
-                tooltip_style=TOOLTIP_STYLE,
-            )
-            marks.append(line)
-            return line
-
-        group = Groups.CASE
-        # remove any existing marks for a selected movement.
-        if group in chart.added_marks_groups:
-            chart.remove_added_marks(group)
-
-        if case.closed:
-            if typing.TYPE_CHECKING:
-                assert case.conf_chg_pct is not None
-                assert case.end_conf_px is not None
-            style = tooltip_html_style(color=color)
-            s = f"<p {style}>Conf chg: "
-            chg_color = COL_DEC if case.conf_chg_pct < 0 else COL_ADV  # type: ignore
-            chg_style = tooltip_html_style(color=chg_color)
-            s += f"<span {chg_style}>{formatter_percent(case.conf_chg_pct)}</span></p>"
-
-            if case.is_adv:
-                color_area = (
-                    COL_ADV if case.start_conf_px < case.end_conf_px else COL_DEC
-                )
-            else:
-                color_area = (
-                    COL_ADV if case.start_conf_px > case.end_conf_px else COL_DEC
-                )
-
-            mark_chg = bq.Lines(
-                x=[[case.start_conf, case.end_conf]] * 2,
-                y=[[case.start_conf_px] * 2, [case.end_conf_px] * 2],
-                scales=chart.scales,
-                opacities=[0],
-                fill="between",
-                fill_colors=[color_area],
-                fill_opacities=[0.2],
-                tooltip=w.HTML(value=s),
-                tooltip_style=TOOLTIP_STYLE,
-            )
-            chart.add_marks([mark_chg], group, under=True)
-
-        def handler_start(mark: bq.Scatter, event: dict):
-            mark.tooltip.value = self.get_move_html(case)
-
-        f([case.start], [case.start_px], color, mark.marker, handler_start)
-
-        handler = self._handler_not_start
-        f([case.start_conf], [case.start_conf_px], color, "circle", handler)
-        if case.closed:
-            f([case.end], [case.end_px], color, "cross", handler)
-            f([case.end_conf], [case.end_conf_px], color, "square", handler)
-
-        color_break, color_limit = "white", "slategray"
-        if not case.by_break:
-            color_break, color_limit = color_limit, color_break
-        fl(case.line_break, color_break, "dashed", "Break line")
-        fl(case.line_limit, color_limit, "dashed", "Limit line")
-
-        chart.hide_cases()
-        chart.add_marks(marks, group)
 
 
 @dataclass(frozen=True)
@@ -784,12 +643,6 @@ class MovementsAlt(MovementsBase, MovementsSupportChartAnaly):
     """Movements identified over an analysis period by `Trends`."""
 
     cases: Sequence[MovementAlt]
-
-    MARKER_MAP: typing.ClassVar = {
-        "cross": "End",
-        "circle": "Start conf",
-        "square": "End conf",
-    }
 
     def event_to_case(self, mark: bq.Scatter, event: dict) -> MovementAlt:
         # Only for type checker to know move is a Movement as defined on this module.
@@ -799,8 +652,9 @@ class MovementsAlt(MovementsBase, MovementsSupportChartAnaly):
         return case
 
     @staticmethod
-    def get_move_html(case: MovementAlt) -> str:
+    def get_case_html(case: CaseSupportsChartAnaly) -> str:
         """Return html to describe a movement."""
+        case = typing.cast(MovementAlt, case)
         color = "crimson" if case.is_dec else "limegreen"
         style = tooltip_html_style(color=color, line_height=1.3)
         s = f"<p {style}>Start: " + formatter_datetime(case.start)
@@ -829,169 +683,3 @@ class MovementsAlt(MovementsBase, MovementsSupportChartAnaly):
 
         s += "</p"
         return s
-
-    def handler_hover_start(self, mark: bq.marks.Scatter, event: dict):
-        """Handler for hovering on mark representing a movement start."""
-        case = self.event_to_case(mark, event)
-        mark.tooltip.value = self.get_move_html(case)
-
-    def _handler_not_start(self, mark: bq.Scatter, event: dict):
-        """Handler to hover over any scatter mark not representing a movement start.
-
-        Displays tooltip describing what scatter represents and corresponding timestamp.
-        """
-        style = tooltip_html_style(color=mark.colors[0], line_height=1.3)
-        date = ubq.discontinuous_date_to_timestamp(event["data"]["x"])
-        name = self.MARKER_MAP[mark.marker]
-        s = f"<p {style}>{name}: {formatter_datetime(date)}"
-        s += f"<br>{name + ' px'}: {formatter_float(event['data']['y'])}</p"
-        mark.tooltip.value = s
-
-    def handler_hover_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing a movement end."""
-        self._handler_not_start(mark, event)
-
-    def handler_hover_conf_start(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing when movement start confirmed."""
-        self._handler_not_start(mark, event)
-
-    def handler_hover_conf_end(self, mark: bq.Scatter, event: dict):
-        """Handler for hovering on mark representing when movement end confirmed."""
-        self._handler_not_start(mark, event)
-
-    def handler_click_case(self, chart: OHLCTrends, mark: bq.Scatter, event: dict):
-        """Handler for clicking mark representing a movement start.
-
-        Removes all existing scatters from figure.
-
-        Adds scatters and lines for the single trend represented by the
-        clicked scatter point.
-        """
-        case = self.event_to_case(mark, event)
-        color = mark.colors[0]
-
-        marks = []
-
-        def f(*args):
-            marks.append(chart.create_scatter(*args))
-
-        def fl(data: pd.Series, color_: str, line_style: str, desc: str) -> bq.Lines:
-            tooltip_str = f"<p {tooltip_html_style(color=color_)}>{desc}</p>"
-            line = bq.Lines(
-                x=data.index,
-                y=data,
-                scales=chart.scales,
-                colors=[color_],
-                line_style=line_style,
-                tooltip=w.HTML(value=tooltip_str),
-                tooltip_style=TOOLTIP_STYLE,
-            )
-            marks.append(line)
-            return line
-
-        group = Groups.CASE
-        # remove any existing marks for a selected case.
-        if group in chart.added_marks_groups:
-            chart.remove_added_marks(group)
-
-        if case.closed:
-            if typing.TYPE_CHECKING:
-                assert case.conf_chg_pct is not None
-                assert case.end_conf_px is not None
-            style = tooltip_html_style(color=color)
-            s = f"<p {style}>Conf chg: "
-            chg_color = COL_DEC if case.conf_chg_pct < 0 else COL_ADV  # type: ignore
-            chg_style = tooltip_html_style(color=chg_color)
-            s += f"<span {chg_style}>{formatter_percent(case.conf_chg_pct)}</span></p>"
-
-            if case.is_adv:
-                color_area = (
-                    COL_ADV if case.start_conf_px < case.end_conf_px else COL_DEC
-                )
-            else:
-                color_area = (
-                    COL_ADV if case.start_conf_px > case.end_conf_px else COL_DEC
-                )
-
-            mark_chg = bq.Lines(
-                x=[[case.start_conf, case.end_conf]] * 2,
-                y=[[case.start_conf_px] * 2, [case.end_conf_px] * 2],
-                scales=chart.scales,
-                opacities=[0],
-                fill="between",
-                fill_colors=[color_area],
-                fill_opacities=[0.2],
-                tooltip=w.HTML(value=s),
-                tooltip_style=TOOLTIP_STYLE,
-            )
-            chart.add_marks([mark_chg], group, under=True)
-
-        def handler_start(mark: bq.Scatter, event: dict):
-            mark.tooltip.value = self.get_move_html(case)
-
-        f([case.start], [case.start_px], color, mark.marker, handler_start)
-
-        handler = self._handler_not_start
-        f([case.start_conf], [case.start_conf_px], color, "circle", handler)
-        if case.closed:
-            f([case.end], [case.end_px], color, "cross", handler)
-            f([case.end_conf], [case.end_conf_px], color, "square", handler)
-
-        fl(case.sel, color, "dashed", "Start Establishment Line")
-        fl(case.start_conf_line, color, "dash_dotted", "Confirmed Start Line")
-        fl(
-            case.end_line_consol,
-            "white" if case.by_consol else "slategray",
-            "dashed",
-            "Conf End Line by Consolidation",
-        )
-        end_line_rvr_col = "white" if case.by_rvr_and_by_pct else "slategray"
-        end_line_rvr = fl(
-            case.end_line_rvr,
-            end_line_rvr_col,
-            "dash_dotted",
-            "Conf End Line by Reversal",
-        )
-
-        # create labels for end_line_rvr
-        rvr_arr = case.rvr_arr
-        xs, ys, texts = [end_line_rvr.x[0]], [end_line_rvr.y[0]], [str(rvr_arr[0])]
-        for i, (x, y, rvr_) in enumerate(
-            zip(end_line_rvr.x[1:], end_line_rvr.y[1:], rvr_arr[1:])
-        ):
-            if rvr_ == rvr_arr[i]:
-                continue
-            xs.append(x)
-            ys.append(y)
-            texts.append(str(rvr_))
-        label_mark = bq.Label(
-            scales=chart.scales,
-            x=xs,
-            y=ys,
-            text=texts,
-            colors=[end_line_rvr_col],
-            default_size=11,
-            y_offset=15 if case.is_adv else -15,
-            enable_move=True,
-            restrict_y=True,
-            visible=False,
-        )
-        marks.append(label_mark)
-
-        def end_line_rvr_handler(*_):
-            label_mark.visible = not label_mark.visible
-
-        end_line_rvr.on_element_click(end_line_rvr_handler)
-
-        if case.end_line_rvr_opp is not None:
-            fl(
-                case.end_line_rvr_opp,
-                "white",
-                "dash_dotted",
-                "Conf End Line by Opposing Movement",
-            )
-        if case.eel is not None:
-            fl(case.eel, color, "dotted", "End Establishment Line")
-
-        chart.hide_cases()
-        chart.add_marks(marks, group)
